requirements.txt
Django==3.2.8
psycopg2-binary==2.9.1
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
python-decouple==3.4
pandas==1.3.3
psycopg2-binary==2.9.1
